Move JS registry to public header, slight rename
