+++ src/apps/js_generic/js_generic_base.cpp
@@ -17,6 +17,7 @@
 #include "ccf/js/extensions/ccf/rpc.h"
 #include "ccf/js/extensions/console.h"
 #include "ccf/js/extensions/math/random.h"
+#include "ccf/js/interpreter_cache_interface.h"
 #include "ccf/js/modules.h"
 #include "ccf/js/named_auth_policies.h"
 #include "ccf/node/host_processes_interface.h"
@@ -25,7 +26,6 @@
 #include "ccf/version.h"
 #include "enclave/enclave_time.h"
 #include "js/global_class_ids.h"
-#include "js/interpreter_cache_interface.h"
 #include "service/tables/endpoints.h"
 
 #include <memory>
