+++ src/js/interpreter_cache.h
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "./interpreter_cache_interface.h"
+#include "ccf/js/interpreter_cache_interface.h"
 #include "ccf/pal/locking.h"
 #include "ds/lru.h"
